+++ devnotes/toolkit/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Toolkit CHANGELOG
 
+## 2026-09-19 (98차)
+- lead_decel/ 추가: events.py, needed_decel.py, gating_eval.py (리드 감속 게이팅 후보 평가, README 참고). gating_eval.py의 별칭 B(G2T)가 98차 채택안. 97차 스크립트 5개는 변경 없음
+
 ## 2026-09-19 (97차)
 - lead_decel/ 추가: parse_lead_log.py, merge_lead_series.py, counterfactual.py, mpc_replica.py, closed_loop.py (선행차 감속 과민 반응 분석용, README 참고)